usb: add standalone dfu class implementation, use it in embassy-usb-dfu.
