+++ embassy-usb/Cargo.toml
@@ -62,11 +62,13 @@ embassy-futures = { version = "0.1.2", path = "../embassy-futures" }
 embassy-usb-driver = { version = "0.2.0", path = "../embassy-usb-driver" }
 embassy-sync = { version = "0.7.2", path = "../embassy-sync" }
 embassy-net-driver-channel = { version = "0.3.2", path = "../embassy-net-driver-channel" }
+embassy-time = { version = "0.5.0", path = "../embassy-time" }
 
 defmt = { version = "1", optional = true }
 log = { version = "0.4.14", optional = true }
 heapless = "0.8"
 embedded-io-async = { version = "0.7.0" }
+bitflags = "2.4.1"
 
 # for HID
 usbd-hid = { version = "0.9.0", optional = true }
