+++ embassy-usb/src/class/mod.rs
@@ -2,6 +2,7 @@
 pub mod cdc_acm;
 pub mod cdc_ncm;
 pub mod cmsis_dap_v2;
+pub mod dfu;
 pub mod hid;
 pub mod midi;
 pub mod uac1;
